Bosse un peu STP J'ai qq soucis pour le basicEntity car mes entité ont besoin aussi d'une position x y et aussi un angle

+++ src/base/engine/entities/others/outputs/IOrigin.java
@@ -8,7 +8,7 @@ package base.engine.entities.others.outputs;
  */
 public interface IOrigin {
 	
-	public int getX();
-	public int getY();
+	public float getX();
+	public float getY();
 
 }
